Support SVG <a> as argument to Origin.from() https://bugs.webkit.org/show_bug.cgi?id=315947
Reviewed by Taher Ali.

MathML <a> will have to wait until we support MathMLAnchorElement.

Upstream commit: https://github.com/web-platform-tests/wpt/commit/34637df05a42cefd99ecc38e6602f7f64e4c1648

Canonical link: https://commits.webkit.org/314287@main
