+++ LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/WEB_FEATURES.yml
@@ -1,0 +1,3 @@
+features:
+- name: origin
+  files: "**"
