+++ LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-from-hyperlinkelementutils.window.html
