+++ LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/w3c-import.log
@@ -14,11 +14,12 @@ Property values requiring vendor prefixes:
 None
 ------------------------------------------------------------------------
 List of files:
+/LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/WEB_FEATURES.yml
 /LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/file-scheme-origin.tentative.any.js
 /LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-comparison.any.js
 /LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-from-extendablemessageevent.any.js
 /LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-from-global.any.js
-/LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-from-htmlhyperlinkelementutils.window.js
+/LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-from-hyperlinkelementutils.window.js
 /LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-from-location.window.js
 /LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-from-messageevent-opaque.window.js
 /LayoutTests/imported/w3c/web-platform-tests/html/browsers/origin/api/origin-from-messageevent.window.js
